Added extension method MapSignalRHubWithOptions to utilize setting it… (#195)
* Added extension method MapSignalRHubWithOptions to utilize setting its properties through configuration

We would change some of signalR default properties, for example max size of message. With this method, it can be set in app configuration instead changing it in code directly.

HttpConnectionDispatcherOptions has default values, if we change some in configuration, it accepts it. In hub configuration we set this as option object, instead of mapping properties.

* update version